MKJD_BE/requirements-core.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
alembic
pydantic-settings
python-dotenv
psycopg2-binary
httpx
google-auth
google-auth-oauthlib
google-api-python-client
